+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 ### Internal
 
 * Removed `Table#Table()`, `Table#addEmptyRow()`, `Table#addEmptyRows()`, `Table#add(Object...)`, `Table#pivot(long,long,PivotType)` and `Table#createnative()`.
+* Renaming `io.realm.internal.SharedRealm` to `io.realm.internal.OsSharedRealm`.
 
 
 ## 3.4.0 (2017-06-22)